Fix issue where connections are aborted prematurely.
If the event loop is blocked or really busy, it's possible that the timeout
callback fires even though the underlying socket is connected. This is because
the I/O phase of the event loop hasn't yet had a chance to process the
connection event.

This change instead allows the event loop to go through all of its phases before
aborting so that I/O events have an opportunity to inform got that connections,
lookups, and other I/O things actually did process.